revert buggy bugfix
The zsh bugfix introduced other bugs.  I think the Edit.results
field needs more thinking in the ambiguous cases.  It should probably
contain some notion of the full transform of how the input symbol
arrived at becoming the speakable version that talon chose, and the
consumer of the Edit will be able to figure out what the right decision
is for its own use case.

In the mean time, I'll hack zsh.py to be a bit smarter and that should
handle 80% of use cases for zsh.